Adds rt-* options for install
R2DT packages ribovore and wants it to be as lightweight as possible.
Only ribotyper is used, so blast is unnecessary. So is the taxonomy subdir.
With the rt-* builds, blast is not downloaded and the taxonomy subdir is
removed.

+++ install.sh
@@ -26,28 +26,66 @@ RVERSION="ribovore-$VERSION"
 
 # set defaults
 INPUTSYSTEM="?"
+RIBOTYPERONLY="?"
 
 ########################
 # Validate correct usage
 ########################
 # make sure correct number of cmdline arguments were used, exit if not
 if [ "$#" -ne 1 ]; then
-   echo "Usage: $0 <\"linux\" or \"macosx-intel\" or \"macosx-silicon\">"
+   echo "Usage:   $0 <os>"
+   echo ""
+   echo "valid options for <os> are:"
+   echo "  \"linux\":             full install for linux"
+   echo "  \"macosx-intel\":      full install for macosx-intel"
+   echo "  \"macosx-silicon\":    full install for macosx-silicon"
+   echo "  \"rt-linux\":          ribotyper only for linux"
+   echo "  \"rt-macosx-intel\":   ribotyper only for macosx-intel"
+   echo "  \"rt-macosx-silicon\": ribotyper only for macosx-silicon"
+   echo ""
+   echo "Example: $0 linux"
+   echo ""
    exit 1
 fi
 
 # make sure 1st argument is either "linux" or "macosx"
 if [ "$1" = "linux" ]; then
     INPUTSYSTEM="linux";
+    RIBOTYPERONLY="no"
 fi
 if [ "$1" = "macosx-intel" ]; then
     INPUTSYSTEM="macosx-intel";
+    RIBOTYPERONLY="no"
 fi
 if [ "$1" = "macosx-silicon" ]; then
     INPUTSYSTEM="macosx-silicon";
+    RIBOTYPERONLY="no"
+fi
+if [ "$1" = "rt-linux" ]; then
+    INPUTSYSTEM="linux";
+    RIBOTYPERONLY="yes"
+fi
+if [ "$1" = "rt-macosx-intel" ]; then
+    INPUTSYSTEM="macosx-intel";
+    RIBOTYPERONLY="yes"
+fi
+if [ "$1" = "rt-macosx-silicon" ]; then
+    INPUTSYSTEM="macosx-silicon";
+    RIBOTYPERONLY="yes"
 fi
 if [ "$INPUTSYSTEM" = "?" ]; then 
-   echo "Usage: $0 <\"linux\" or \"macosx-intel\" or \"macosx-silicon\">"
+   echo "Usage:   $0 <os>"
+   echo ""
+   echo "valid options for <os> are:"
+   echo "  \"linux\":             full install for linux"
+   echo "  \"macosx-intel\":      full install for macosx-intel"
+   echo "  \"macosx-silicon\":    full install for macosx-silicon"
+   echo "  \"rt-linux\":          ribotyper only for linux"
+   echo "  \"rt-macosx-intel\":   ribotyper only for macosx-intel"
+   echo "  \"rt-macosx-silicon\": ribotyper only for macosx-silicon"
+   echo ""
+   echo "Example: $0 linux"
+   echo ""
    exit 1
 fi
 ########################################################
@@ -80,18 +118,9 @@ curl -k -L -o $RVERSION.zip https://github.com/ncbi/ribovore/archive/$RVERSION.z
 #rm -rf .git
 #cd ..
 # ----------------------------------------------------------------------------
-
-# rRNA_sensor
-echo "Downloading rRNA_sensor ... "
-curl -k -L -o rRNA_sensor-$RVERSION.zip https://github.com/aaschaffer/rRNA_sensor/archive/$RVERSION.zip; unzip rRNA_sensor-$RVERSION.zip; mv rRNA_sensor-$RVERSION rRNA_sensor; rm rRNA_sensor-$RVERSION.zip
-# to checkout a specific branch, comment out above curl and uncomment block below
-# ----------------------------------------------------------------------------
-#git clone https://github.com/aaschaffer/rRNA_sensor.git rRNA_sensor
-#cd rRNA_sensor
-#git checkout release-0.14
-#rm -rf .git
-#cd ..
-# ----------------------------------------------------------------------------
+if [ "$RIBOTYPERONLY" = "yes" ]; then
+    rm -rf ribovore/taxonomy
+fi
 
 # sequip
 echo "Downloading sequip ... "
@@ -104,6 +133,20 @@ curl -k -L -o sequip-$RVERSION.zip https://github.com/nawrockie/sequip/archive/$
 #rm -rf .git
 #cd ..
 # ----------------------------------------------------------------------------
+
+if [ "$RIBOTYPERONLY" != "yes" ]; then
+    # rRNA_sensor
+    echo "Downloading rRNA_sensor ... "
+    curl -k -L -o rRNA_sensor-$RVERSION.zip https://github.com/aaschaffer/rRNA_sensor/archive/$RVERSION.zip; unzip rRNA_sensor-$RVERSION.zip; mv rRNA_sensor-$RVERSION rRNA_sensor; rm rRNA_sensor-$RVERSION.zip
+    # to checkout a specific branch, comment out above curl and uncomment block below
+    # ----------------------------------------------------------------------------
+    #git clone https://github.com/aaschaffer/rRNA_sensor.git rRNA_sensor
+    #cd rRNA_sensor
+    #git checkout release-0.14
+    #rm -rf .git
+    #cd ..
+    # ----------------------------------------------------------------------------
+fi
 
 # download two infernal binary distributions
 # the first is a recent release, we'll use most programs from this release,
@@ -118,34 +161,36 @@ if [ "$INPUTSYSTEM" = "linux" ]; then
     curl -k -L -o infernal.tar.gz http://eddylab.org/infernal/infernal-$IVERSION-linux-intel-gcc.tar.gz
     tar xfz infernal.tar.gz
     rm infernal.tar.gz
-    echo "Downloading Infernal version $IESLCLUSTERVERSION for Linux"
-    curl -k -L -o infernal2.tar.gz http://eddylab.org/infernal/infernal-$IESLCLUSTERVERSION-linux-intel-gcc.tar.gz
-    tar xfz infernal2.tar.gz
-    rm infernal2.tar.gz
+    mv infernal-$IVERSION-linux-intel-gcc infernal
+    if [ "$RIBOTYPERONLY" != "yes" ]; then
+        echo "Downloading Infernal version $IESLCLUSTERVERSION for Linux"
+        curl -k -L -o infernal2.tar.gz http://eddylab.org/infernal/infernal-$IESLCLUSTERVERSION-linux-intel-gcc.tar.gz
+        tar xfz infernal2.tar.gz
+        rm infernal2.tar.gz
+        mv infernal-$IESLCLUSTERVERSION-linux-intel-gcc/binaries/esl-cluster infernal/binaries/
+        rm -rf infernal-$IESLCLUSTERVERSION-linux-intel-gcc
+    fi
 elif [ "$INPUTSYSTEM" = "macosx-intel" ]; then       
     echo "Downloading Infernal version $IVERSION for Mac/OSX-intel"
     curl -k -L -o infernal.tar.gz http://eddylab.org/infernal/infernal-$IVERSION-macosx-intel.tar.gz
     tar xfz infernal.tar.gz
     rm infernal.tar.gz
-    echo "Downloading Infernal version $IESLCLUSTERVERSION for Mac/OSX-intel"
-    curl -k -L -o infernal2.tar.gz http://eddylab.org/infernal/infernal-$IESLCLUSTERVERSION-macosx-intel.tar.gz
-    tar xfz infernal2.tar.gz
-    rm infernal2.tar.gz
+    mv infernal-$IVERSION-macosx-intel infernal
+    if [ "$RIBOTYPERONLY" != "yes" ]; then
+        echo "Downloading Infernal version $IESLCLUSTERVERSION for Mac/OSX-intel"
+        curl -k -L -o infernal2.tar.gz http://eddylab.org/infernal/infernal-$IESLCLUSTERVERSION-macosx-intel.tar.gz
+        tar xfz infernal2.tar.gz
+        rm infernal2.tar.gz
+        mv infernal-$IESLCLUSTERVERSION-macosx-intel/binaries/esl-cluster infernal/binaries/
+        rm -rf infernal-$IESLCLUSTERVERSION-macosx-intel
+    fi
 else
     echo "Downloading Infernal version $IVERSION for Mac/OSX-silicon"
     curl -k -L -o infernal.tar.gz http://eddylab.org/infernal/infernal-$IVERSION-macosx-intel.tar.gz
     # esl-cluster doesn't exist in an infernal release that builds on silicon
-fi
-if [ "$INPUTSYSTEM" = "linux" ]; then
-    mv infernal-$IVERSION-linux-intel-gcc infernal
-    mv infernal-$IESLCLUSTERVERSION-linux-intel-gcc/binaries/esl-cluster infernal/binaries/
-    rm -rf infernal-$IESLCLUSTERVERSION-linux-intel-gcc
-elif [ "$INPUTSYSTEM" = "macosx-intel" ]; then       
+    tar xfz infernal.tar.gz
+    rm infernal.tar.gz
     mv infernal-$IVERSION-macosx-intel infernal
-    mv infernal-$IESLCLUSTERVERSION-macosx-intel/binaries/esl-cluster infernal/binaries/
-    rm -rf infernal-$IESLCLUSTERVERSION-macosx-intel
-else
-    mv infernal-$IVERSION-macosx-silicon infernal
 fi
 # ----- infernal block 1 end -----
 
@@ -153,18 +198,15 @@ fi
 # (maybe because the binaries aren't working for you for some reason)
 # comment out 'infernal block 1' above and 
 # uncomment 'infernal block 2' below
+#
 # ----- infernal block 2 start  -----
 #echo "Downloading Infernal version $IVERSION src distribution"
 #curl -k -L -o infernal.tar.gz http://eddylab.org/infernal/infernal-$IVERSION.tar.gz
-#echo "Downloading Infernal version $IESLCLUSTERVERSION src distribution"
-#curl -k -L -o infernal2.tar.gz http://eddylab.org/infernal/infernal-$IESLCLUSTERVERSION.tar.gz
 #tar xfz infernal.tar.gz
-#tar xfz infernal2.tar.gz
 #rm infernal.tar.gz
-#rm infernal2.tar.gz
+#
 #echo "Building Infernal ... "
 #mv infernal-$IVERSION infernal
-#mv infernal-$IESLCLUSTERVERSION infernal2
 #cd infernal
 #mkdir binaries
 #sh ./configure --bindir=$PWD/binaries --prefix=$PWD
@@ -172,41 +214,39 @@ fi
 #make install
 #(cd easel/miniapps; make install)
 #cd ..
-#cd infernal2
-#sh ./configure
-#make
-#mv easel/miniapps/esl-cluster ../infernal/binaries/
-#cd ..
-#rm -rf infernal2
-#echo "Finished building Infernal "
+#echo "Finished building Infernal version $IESLCLUSTERVERSION"
+#
 # ----- infernal block 2 end -----
-echo "------------------------------------------------"
-
+#
+################################
+    
 # download blast binaries
-if [ "$INPUTSYSTEM" = "linux" ]; then
-echo "Downloading BLAST version $BVERSION for Linux"
-curl -k -L -o blast.tar.gz https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/$BVERSION/ncbi-blast-$BVERSION+-x64-linux.tar.gz
-else 
-echo "Downloading BLAST version $BVERSION for Mac/OSX"
-echo "curl -k -L -o blast.tar.gz https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/$BVERSION/ncbi-blast-$BVERSION+-x64-macosx.tar.gz"
-curl -k -L -o blast.tar.gz https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/$BVERSION/ncbi-blast-$BVERSION+-x64-macosx.tar.gz
+if [ "$RIBOTYPERONLY" != "yes" ]; then
+    if [ "$INPUTSYSTEM" = "linux" ]; then
+        echo "Downloading BLAST version $BVERSION for Linux"
+        curl -k -L -o blast.tar.gz https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/$BVERSION/ncbi-blast-$BVERSION+-x64-linux.tar.gz
+    else 
+        echo "Downloading BLAST version $BVERSION for Mac/OSX"
+        echo "curl -k -L -o blast.tar.gz https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/$BVERSION/ncbi-blast-$BVERSION+-x64-macosx.tar.gz"
+        curl -k -L -o blast.tar.gz https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/$BVERSION/ncbi-blast-$BVERSION+-x64-macosx.tar.gz
+    fi
+    tar xfz blast.tar.gz
+    rm blast.tar.gz
+    mv ncbi-blast-$BVERSION+ ncbi-blast
+    echo "------------------------------------------------"
+    
+    # if linux, download vecscreen_plus_taxonomy
+    if [ "$INPUTSYSTEM" = "linux" ]; then
+        curl -k -L -o vecscreen_plus_taxonomy-$RVERSION.zip https://github.com/aaschaffer/vecscreen_plus_taxonomy/archive/$RVERSION.zip; 
+        unzip vecscreen_plus_taxonomy-$RVERSION.zip; 
+        mv vecscreen_plus_taxonomy-$RVERSION vecscreen_plus_taxonomy
+        rm vecscreen_plus_taxonomy-$RVERSION.zip
+        (cd vecscreen_plus_taxonomy/scripts; gunzip srcchk.gz; gunzip vecscreen.gz;)
+    else 
+        echo "Not installing vecscreen_plus_taxonomy (only available for Linux)"
+    fi
+    echo "------------------------------------------------"
 fi
-tar xfz blast.tar.gz
-rm blast.tar.gz
-mv ncbi-blast-$BVERSION+ ncbi-blast
-echo "------------------------------------------------"
-
-# if linux, download vecscreen_plus_taxonomy
-if [ "$INPUTSYSTEM" = "linux" ]; then
-curl -k -L -o vecscreen_plus_taxonomy-$RVERSION.zip https://github.com/aaschaffer/vecscreen_plus_taxonomy/archive/$RVERSION.zip; 
-unzip vecscreen_plus_taxonomy-$RVERSION.zip; 
-mv vecscreen_plus_taxonomy-$RVERSION vecscreen_plus_taxonomy
-rm vecscreen_plus_taxonomy-$RVERSION.zip
-(cd vecscreen_plus_taxonomy/scripts; gunzip srcchk.gz; gunzip vecscreen.gz;)
-else 
-echo "Not installing vecscreen_plus_taxonomy (only available for Linux)"
-fi
-echo "------------------------------------------------"
 
 ###############################################
 # Message about setting environment variables
@@ -224,18 +264,23 @@ echo "export RIBOSCRIPTSDIR=\"\$RIBOINSTALLDIR/ribovore\""
 echo "export RIBOINFERNALDIR=\"\$RIBOINSTALLDIR/infernal/binaries\""
 echo "export RIBOEASELDIR=\"\$RIBOINSTALLDIR/infernal/binaries\""
 echo "export RIBOSEQUIPDIR=\"\$RIBOINSTALLDIR/sequip\""
-echo "export RIBOBLASTDIR=\"\$RIBOINSTALLDIR/ncbi-blast/bin\""
-echo "export RIBOTIMEDIR=/usr/bin"
-echo "export RRNASENSORDIR=\"\$RIBOINSTALLDIR/rRNA_sensor\""
-if [ "$INPUTSYSTEM" = "linux" ]; then
-echo "export VECPLUSDIR=\"\$RIBOINSTALLDIR/vecscreen_plus_taxonomy\""
-echo "export BLASTDB=\"\$VECPLUSDIR/univec-files\":\"\$RRNASENSORDIR\":\"\$BLASTDB\""
-echo "export PERL5LIB=\"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$VECPLUSDIR\":\"\$PERL5LIB\""
+echo "export RIBOTIMEDIR=\"/usr/bin\""
+if [ "$RIBOTYPERONLY" != "yes" ]; then
+    echo "export RIBOBLASTDIR=\"\$RIBOINSTALLDIR/ncbi-blast/bin\""
+    echo "export RRNASENSORDIR=\"\$RIBOINSTALLDIR/rRNA_sensor\""
+    if [ "$INPUTSYSTEM" = "linux" ]; then
+        echo "export VECPLUSDIR=\"\$RIBOINSTALLDIR/vecscreen_plus_taxonomy\""
+        echo "export BLASTDB=\"\$VECPLUSDIR/univec-files\":\"\$RRNASENSORDIR\":\"\$BLASTDB\""
+        echo "export PERL5LIB=\"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$VECPLUSDIR\":\"\$PERL5LIB\""
+    else
+        echo "export BLASTDB=\"\$RRNASENSORDIR\":\"\$BLASTDB\""
+        echo "export PERL5LIB=\"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$PERL5LIB\""
+    fi
+    echo "export PATH=\"\$RIBOSCRIPTSDIR\":\"\$RIBOBLASTDIR\":\"\$RRNASENSORDIR\":\"\$PATH\""
 else
-echo "export BLASTDB=\"\$RRNASENSORDIR\":\"\$BLASTDB\""
-echo "export PERL5LIB=\"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$PERL5LIB\""
+    echo "export PERL5LIB=\"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$PERL5LIB\""
+    echo "export PATH=\"\$RIBOSCRIPTSDIR\":\"\$PATH\""
 fi
-echo "export PATH=\"\$RIBOSCRIPTSDIR\":\"\$RIBOBLASTDIR\":\"\$RRNASENSORDIR\":\"\$PATH\""
 echo ""
 echo "After adding the export lines to your .bashrc or .zshrc file, source that file"
 echo "to update your current environment with the command:"
@@ -256,18 +301,23 @@ echo "setenv RIBOSCRIPTSDIR \"\$RIBOINSTALLDIR/ribovore\""
 echo "setenv RIBOINFERNALDIR \"\$RIBOINSTALLDIR/infernal/binaries\""
 echo "setenv RIBOEASELDIR \"\$RIBOINSTALLDIR/infernal/binaries\""
 echo "setenv RIBOSEQUIPDIR \"\$RIBOINSTALLDIR/sequip\""
-echo "setenv RIBOBLASTDIR \"\$RIBOINSTALLDIR/ncbi-blast/bin\""
-echo "setenv RIBOTIMEDIR /usr/bin"
-echo "setenv RRNASENSORDIR \"\$RIBOINSTALLDIR/rRNA_sensor\""
-if [ "$INPUTSYSTEM" = "linux" ]; then
-echo "setenv VECPLUSDIR \"\$RIBOINSTALLDIR/vecscreen_plus_taxonomy\""
-echo "setenv BLASTDB \"\$VECPLUSDIR/univec-files\":\"\$RRNASENSORDIR\":\"\$BLASTDB\""
-echo "setenv PERL5LIB \"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$VECPLUSDIR\":\"\$PERL5LIB\""
+echo "setenv RIBOTIMEDIR \"/usr/bin\""
+if [ "$RIBOTYPERONLY" != "yes" ]; then
+    echo "setenv RIBOBLASTDIR \"\$RIBOINSTALLDIR/ncbi-blast/bin\""
+    echo "setenv RRNASENSORDIR \"\$RIBOINSTALLDIR/rRNA_sensor\""
+    if [ "$INPUTSYSTEM" = "linux" ]; then
+        echo "setenv VECPLUSDIR \"\$RIBOINSTALLDIR/vecscreen_plus_taxonomy\""
+        echo "setenv BLASTDB \"\$VECPLUSDIR/univec-files\":\"\$RRNASENSORDIR\":\"\$BLASTDB\""
+        echo "setenv PERL5LIB \"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$VECPLUSDIR\":\"\$PERL5LIB\""
+    else
+        echo "setenv BLASTDB \"\$RRNASENSORDIR\":\"\$BLASTDB\""
+        echo "setenv PERL5LIB \"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$PERL5LIB\""
+    fi
+    echo "setenv PATH \"\$RIBOSCRIPTSDIR\":\"\$RIBOBLASTDIR\":\"\$RRNASENSORDIR\":\"\$PATH\""
 else
-echo "setenv BLASTDB \"\$RRNASENSORDIR\":\"\$BLASTDB\""
-echo "setenv PERL5LIB \"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$PERL5LIB\""
+    echo "setenv PERL5LIB \"\$RIBOSCRIPTSDIR\":\"\$RIBOSEQUIPDIR\":\"\$PERL5LIB\""
+    echo "setenv PATH \"\$RIBOSCRIPTSDIR\":\"\$PATH\""
 fi
-echo "setenv PATH \"\$RIBOSCRIPTSDIR\":\"\$RIBOBLASTDIR\":\"\$RRNASENSORDIR\":\"\$PATH\""
 echo ""
 echo "After adding the setenv lines to your .cshrc file, source that file"
 echo "to update your current environment with the command:"
